% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colombian.R
\docType{data}
\name{colombian}
\alias{colombian}
\title{Colombian}
\format{
A data frame with seven variables:

\describe{
\item{\code{share}}{level intermediate input's revenue share}
\item{\code{id}}{firm id}
\item{\code{year}}{the last two digits of a year between 1981 and 1991}
\item{\code{RGO}}{level of real gross output with base year 1981}
\item{\code{L}}{labor in level employee years}
\item{\code{K}}{level of real capital stock with base year 1981}
\item{\code{RI}}{level of real intermediate inputs with base year 1981}
}

For a complete listing of data, see \url{https://www.journals.uchicago.edu/doi/suppl/10.1086/707736}
}
\usage{
colombian
}
\description{
Plant-level production data for the food products industry (International Standard Industrial Classification code 311) in Colombia.
}
\keyword{datasets}
