% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrflex.R
\name{gnrflex}
\alias{gnrflex}
\title{Estimate flexible input elasticity: Gandhi, Navarro, Rivers (GNR) share regression}
\usage{
gnrflex(output, fixed, flex, share, id, time, degree, control)
}
\arguments{
\item{output}{a numeric vector of level gross output}

\item{fixed}{a numeric matrix of level fixed inputs}

\item{flex}{a numeric vector of level flexible input}

\item{share}{a numeric vector of the level intermediate input's revenue share}

\item{id}{a numeric vector of firm ids}

\item{time}{a numeric vector of time}

\item{degree}{degree of share regression polynomial}

\item{control}{an optional list of convergence settings. See \code{gnrflex.control} for listing.}
}
\value{
a list of class "gnrflex" containing three elements:

\code{elas}: a list containing six elements describing the share regression:

\itemize{
\item{\code{flex_in_elas}}{: a numeric vector of the estimated flexible input elasticity for each observation}
\item{\code{coef}}{: a numeric vector of the coefficients of the estimator scaled by a constant (equation (21))}
\item{\code{residuals}}{: a numeric vector of the residuals}
\item{\code{SSR}}{: sum of squared residuals}
\item{\code{iterations}}{: number of iterations performed}
\item{\code{convergence}}{: boolean indicating whether convergence was achieved}
}

\code{arg}: a list containing seven elements to be passed to the second stage function \code{gnriv}
\itemize{
\item{\code{input}}{: a numeric matrix (S3 'poly') of the polynomial expansion of all inputs}
\item{\code{input_degree}}{: a numeric matrix corresponding to \code{input} denoting each vector's degree}
\item{\code{big_Y}}{: a numeric vector of persistent productivity minus the constant of integration (equation (16) in Gandhi, Navarro, and Rivers (2020))}
\item{\code{D_coef}}{: a numeric vector equalling \code{coef} divided by an estimate of the constant}
\item{\code{id}}{: a numeric vector of the firm ids}
\item{\code{time}}{: a numeric vector of time}
\item{\code{degree}}{: the degree of the share regression}
}

\code{control}: the list of convergence control parameters. See \code{gnrflex.control}
}
\description{
The \code{gnrflex} function implements the first stage (share regression) of the GNR production function estimation routine, nonparametrically identifying the flexible input elasticity of the production function. This function is called in the main wrapper function \code{gnrprod}. It currently supports only one flexible input.

For details, see Gandhi, Navarro, and Rivers (2020).
}
