% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnriv.control.R
\name{gnriv.control}
\alias{gnriv.control}
\title{Control iterations in second stage of GNR}
\usage{
gnriv.control(maxit = 100, reltol = 1e-08)
}
\arguments{
\item{maxit}{Maximum number of iterations. Defaults to 100.}

\item{reltol}{Relative convergence tolerance. Defaults to 1e-8.}
}
\value{
a list containing two elements: \code{maxit} and \code{reltol}
}
\description{
Allows the user to modify convergence parameters of Gauss Newton algorithm used in the \code{gnriv} function
}
