% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{print.gnriv}
\alias{print.gnriv}
\title{Printing second stage estimates}
\usage{
\method{print}{gnriv}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class 'gnriv'.}

\item{digits}{the number of significant figures to use for printing.}

\item{...}{currently not used.}
}
\description{
Print estimates of the fixed input elasticities and productivity from \code{\link[gnrprod]{gnriv}}
}
