% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnriv.control.R
\name{gnriv.control}
\alias{gnriv.control}
\title{Control iterations in second stage of GNR}
\usage{
gnriv.control(
  method = "BFGS",
  trace,
  fnscale,
  parscale,
  ndeps,
  maxit,
  abstol,
  reltol,
  alpha,
  beta,
  gamma,
  REPORT,
  type,
  lmm,
  factr,
  pgtol,
  tmax,
  temp
)
}
\arguments{
\item{method}{temp}

\item{trace}{temp}

\item{fnscale}{temp}

\item{parscale}{temp}

\item{ndeps}{temp}

\item{maxit}{temp}

\item{abstol}{temp}

\item{reltol}{temp}

\item{alpha}{temp}

\item{beta}{temp}

\item{gamma}{temp}

\item{REPORT}{temp}

\item{type}{temp}

\item{lmm}{temp}

\item{factr}{temp}

\item{pgtol}{temp}

\item{tmax}{temp}

\item{temp}{temp}
}
\value{
a list of called control parameters
}
\description{
Allows the user to modify convergence parameters of Gauss Newton algorithm used in the \code{gnriv} function
}
