% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnriv.R
\name{gnriv}
\alias{gnriv}
\title{Estimate fixed input elasticity and total productivity: Gandhi, Navarro, Rivers (GNR) lag instruments}
\usage{
gnriv(object, degree, control)
}
\arguments{
\item{object}{object of class \code{gnrflex}}

\item{degree}{degree of Markov process for persistent productivity}

\item{control}{an optional list of convergence settings. See \code{gnriv.control} for listing.}
}
\value{
a list of class "gnriv" containing six elements:

\code{elas}: a numeric matrix of estimated elasticities of fixed inputs for each observation

\code{productivity}: a numeric vector of estimated total productivity

\code{degree}: degree of Markov process

\code{iterations}: number of iterations performed

\code{convergence}: boolean indicating whether convergence was achieved

\code{control}: the list of convergence control parameters. See \code{gnriv.control}.
}
\description{
The \code{gnriv} function implements the second stage of the GNR production function estimation routine, nonparametrically identifying the fixed input elasticities of the production function and total productivity. This function is called in the main wrapper function \code{gnrprod}. It currently supports only one flexible input.

For details, see Gandhi, Navarro, and Rivers (2020).
}
