% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrprod.R
\name{gnrprod}
\alias{gnrprod}
\title{Estimate the nonparametric model in two stages: the share regression and instrumental variable estimation}
\usage{
gnrprod(
  output,
  fixed,
  flex,
  share,
  in_price = NULL,
  out_price = NULL,
  id,
  time,
  data,
  degree = 2,
  markov_degree = 2,
  firstStageOnly = FALSE,
  fs_control = list(),
  ss_control = list()
)
}
\arguments{
\item{output}{name (character) of variable of level gross output or a numeric vector/dataframe/matrix}

\item{fixed}{name (character or character vector) of variables of level fixed inputs or a numeric vector/dataframe/matrix}

\item{flex}{name (character) of variable of level flexible input or a numeric vector/dataframe/matrix}

\item{share}{name (character) of variable of level flexible input's revenue share or a numeric vector/dataframe/matrix}

\item{in_price}{optional name (character) of variable of common flexible input price or a numeric vector/dataframe/matrix}

\item{out_price}{optional name (character) of variable of common output price or a numeric vector/dataframe/matrix}

\item{id}{name (character) of variable of firm id or a numeric vector/dataframe/matrix}

\item{time}{name (character) of variable of time or a numeric vector/dataframe/matrix}

\item{data}{dataframe containing all variables with names specified by arguments above (left empty if arguments above are vector/dataframe/matrix)}

\item{degree}{degree of share regression polynomial}

\item{markov_degree}{degree of Markovian process for persistent productivity}

\item{firstStageOnly}{boolean for whether return is only first stage results}

\item{fs_control}{list of convergence parameters for first stage}

\item{ss_control}{list of convergence parameters for second stage}
}
\description{
Estimate the nonparametric model in two stages: the share regression and instrumental variable estimation
}
