% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrprod.R
\name{gnrprod}
\alias{gnrprod}
\title{Estimate production functions and productivity: Gandhi, Navarro, and Rivers (2020)}
\usage{
gnrprod(output, fixed, flex, share, in_price = NULL,
               out_price = NULL, id, time, data, degree = 2,
               markov_degree = 2, fs_control = gnrflex.control(),
               ss_control = gnriv.control())
}
\arguments{
\item{output}{name (character) of variable of level gross output or a numeric vector}

\item{fixed}{name (character or character vector) of variables of level fixed inputs or a numeric matrix}

\item{flex}{name (character) of variable of level flexible input or a numeric vector}

\item{share}{name (character) of variable of level intermediate input's revenue share or a numeric vector}

\item{in_price}{optional (required if \code{share} is not specified) name (character) of variable of common flexible input price or a numeric vector}

\item{out_price}{optional (required if \code{share} is not specified) name (character) of variable of common output price or a numeric vector}

\item{id}{name (character) of variable of firm id or a numeric vector}

\item{time}{name (character) of variable of time or a numeric vector}

\item{data}{dataframe containing all variables with names specified by arguments above (left empty if arguments above are vector/matrix)}

\item{degree}{degree of share regression polynomial}

\item{markov_degree}{degree of Markov process for persistent productivity}

\item{fs_control}{an optional list of convergence settings of the first stage. See \code{gnrflex.control} for listing.}

\item{ss_control}{an optional list of convergence settings of the second stage. See \code{gnriv.control} for listing.}
}
\value{
a list of class "gnr" with five elements:
\code{avg_elasticity}: a named numeric vector of the average elasticities of all inputs

\code{data}: a list (dataframe) containing:\code{output}, \code{fixed}, \code{flex}, \code{id}, \code{time}, and \code{share} variables and estimated elasticities for each observation

\code{first_stage}: a list containing six elements describing the share regression (first stage):
\itemize{
\item{\code{coef}}{: a numeric vector of the coefficients of the estimator scaled by a constant (equation (21))}
\item{\code{residuals}}{: a numeric vector of the residuals}
\item{\code{SSR}}{: sum of squared residual}
\item{\code{iterations}}{: number of iterations performed}
\item{\code{convergence}}{: boolean indicating whether convergence was achieved}
\item{\code{control}}{: list of convergence control parameters (see \code{gnrflex.control})}
}

\code{second_stage}: a list containing four elements describing the second stage:
\itemize{
\item{\code{productivity}}{: a numeric vector of the estimated total productivity}
\item{\code{iterations}}{: number of iterations performed}
\item{\code{convergence}}{: boolean indicating whether convergence was achieved}
\item{\code{control}}{: list of convergence control parameters (see \code{gnriv.control})}
}
}
\description{
The \code{gnrprod} function is the front end of the \code{gnrprod} package. It estimates production functions and productivity in two stages: \code{gnrflex} (estimate flexible input elasticity) and \code{gnriv} (estimate fixed input elasticities and productivity). It currently supports only one flexible input.
}
\examples{
require(gnrprod)
data <- colombian
industry_311 <- gnrprod(output = "RGO", fixed = c("L", "K"), flex = "RI",
                        share = "share", id = "id", time = "year",
                        data = data, degree = 3, markov_degree = 3)
}
